Update: Removed export/import and added logging of questionData
I made a quick update to the code. I removed the export/import statements, as they were unnecessary and causing some confusion. Also, I added a log statement to print out the contents of the questionData array to the console.

Added console.error for index bounds checking

I added a console.error statement to check whether the index variable was staying within bounds or going out of bounds. After some testing, I found that it was consistently going out of bounds. To address this issue, I implemented the following fixes:
- First, I added a check to ensure that the index stays within the bounds of our questionData array. This prevents any sneaky index values from causing trouble.
- Next, I ensured that I only increment the index when there are actually more questions available. I don't want to go beyond the last question and end up in no man's land!
- And just in case the index does go out of bounds due to some weird twist of fate, It'll log an error message to the console.